\name{RRHO-package}
\alias{RRHO-package}
\docType{package}
\title{
	Test overlap using the Rank-Rank Hypergeometric test
}
\description{
The package is aimed at inference on the amount of agreement in two sorted lists using the Rank-Rank Hypergeometric Overlap test.
}
\details{
\tabular{ll}{
Package: \tab RRHO\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2013-06-21\cr
License: \tab GPL-2\cr
}
See \code{\link{RRHO}} to get started.
}
\author{
Jason Stein and Jonathan Rosenblatt
Maintainer:  Jonathan Rosenblatt <john.ros.work@gmail.com>
}
\keyword{htest}
\seealso{
	\code{\link{RRHO}}
}
