\name{HNP}
\alias{HNP}
\alias{My}
\alias{Sestan}
\docType{data}
\title{
RRHO comparison data sets
}
\description{
[TODO]
}
\usage{data(lists)}

\format{
  Three data frames: \code{HNP}, \code{My} and \code{Sestan}.
  Each includes gene identifiers and sorting values so that they can be used as inputes to \code{\link{RRHOComparison}}.
}
    
\details{
[TODO]
}
\source{
[TODO]
}
\references{
[TODO]
}
\examples{
data(lists)
## maybe str(HNP) ; plot(HNP) ...
}
\keyword{datasets}
